+++ src/components/RideDetails/index.tsx
@@ -84,12 +84,12 @@ export const RideDetails = ({ ride, user, role, embedded }: Props) => {
             {(isGoing || isGoingAnonymously) && (
               <Button
                 variant="custom"
-                className="flex w-28 items-center justify-center rounded bg-slate-700 p-5 hover:bg-slate-800"
+                className="flex w-32 items-center justify-center rounded bg-slate-700 p-5 hover:bg-slate-800"
                 onClick={openNotes}
               >
                 <div className="flex items-center gap-2">
                   <i className="fa-solid fa-pen-to-square" />
-                  Note
+                  Message
                 </div>
               </Button>
             )}